+++ Enterwell.CI.Changelog.DevOpsExtension/MergeChangelogTask/MergeChangelogTaskV2/tests/L0.ts
@@ -1,15 +1,15 @@
 /**
  * File defining our tests.
  */
-import assert = require("assert");
-import path = require("path");
-import fs = require("fs");
-import ttm = require("azure-pipelines-task-lib/mock-test");
+import assert = require('assert');
+import path = require('path');
+import fs = require('fs');
+import ttm = require('azure-pipelines-task-lib/mock-test');
 
 /**
  * Path to the temporary test folder structure used in our test suite.
  */
-const testRoot = path.join(__dirname, "test_structure");
+const testRoot = path.join(__dirname, 'test_structure');
 
 /**
  * Removes folder on the provided path.
@@ -37,13 +37,13 @@ const removeFolder = (currentPath: string) => {
  */
 const createChanges = () => {
     
-    const changesPath = path.join(testRoot, "changes");
+    const changesPath = path.join(testRoot, 'changes');
     fs.mkdirSync(changesPath);
 
-    fs.writeFileSync(path.join(changesPath, "Added [FE       ] First Change"), "");
-    fs.writeFileSync(path.join(changesPath, "Removed [      API      ] First Deletion"), "");
-    fs.writeFileSync(path.join(changesPath, "Removed [API      ] First Deletion2"), "");
-    fs.writeFileSync(path.join(changesPath, "Removed [API] First Deletion3"), "");
+    fs.writeFileSync(path.join(changesPath, 'Added [FE       ] First Change'), '');
+    fs.writeFileSync(path.join(changesPath, 'Removed [      API      ] First Deletion'), '');
+    fs.writeFileSync(path.join(changesPath, 'Removed [API      ] First Deletion2'), '');
+    fs.writeFileSync(path.join(changesPath, 'Removed [API] First Deletion3'), '');
 }
 
 /**
@@ -62,13 +62,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Migrated from .NET Framework 4.5 to .NET Standard 2.0`;
 
-    fs.writeFileSync(path.join(testRoot, "Changelog.md"), changelogContent);
+    fs.writeFileSync(path.join(testRoot, 'Changelog.md'), changelogContent);
 }
 
 /**
  * Describes a Test Suite (root of our tests).
  */
-describe("Merge Changelog Test Suite", function() {
+describe('Merge Changelog Test Suite', function() {
     this.timeout(60000);
 
     /**
@@ -98,8 +98,8 @@ describe("Merge Changelog Test Suite", function() {
             done();
         }
         catch (error) {
-            console.log("STDERR", tr.stderr);
-            console.log("STDOUT", tr.stdout);
+            console.log('STDERR', tr.stderr);
+            console.log('STDOUT', tr.stdout);
             done(error);
         }
     }
@@ -107,29 +107,29 @@ describe("Merge Changelog Test Suite", function() {
     /**
      * Describes tests with valid inputs.
      */
-    describe("valid inputs", function() {
+    describe('valid inputs', function() {
 
         /**
          * Testing the task when valid inputs are passed but the changelog file does not exist.
          */
-        it("should fail if changelog.md does not exist", (done: Mocha.Done) => {
+        it('should fail if changelog.md does not exist', (done: Mocha.Done) => {
             this.timeout(5000);
     
             // Arrange
             createChanges();
     
             // Act 
-            let tp: string = path.join(__dirname, "L0Valid.js");
+            let tp: string = path.join(__dirname, 'L0Valid.js');
             let tr: ttm.MockTestRunner = new ttm.MockTestRunner(tp);
     
             tr.run();
     
             // Assert
             runValidations(() => {
-             assert(tr.failed, "should have failed");
-             assert.strictEqual(tr.warningIssues.length, 0, "should have no warnings");
-             assert.strictEqual(tr.errorIssues.length, 1, "should have 1 error issue");
-             assert(tr.errorIssues[0].startsWith("Error occurred while reading changelog file"), "should have correct message");
+             assert(tr.failed, 'should have failed');
+             assert.strictEqual(tr.warningIssues.length, 0, 'should have no warnings');
+             assert.strictEqual(tr.errorIssues.length, 1, 'should have 1 error issue');
+             assert(tr.errorIssues[0].startsWith('Error occurred while reading changelog file'), 'should have correct message');
             }, 
             tr, done);
         });
@@ -137,24 +137,24 @@ describe("Merge Changelog Test Suite", function() {
         /**
          * Testing the task when valid inputs are passed but the changes directory does not exist.
          */
-        it("should fail if changes directory does not exist", (done: Mocha.Done) => {
+        it('should fail if changes directory does not exist', (done: Mocha.Done) => {
             this.timeout(5000);
     
             // Arrange
             createChangelogFile();
     
             // Act 
-            let tp: string = path.join(__dirname, "L0Valid.js");
+            let tp: string = path.join(__dirname, 'L0Valid.js');
             let tr: ttm.MockTestRunner = new ttm.MockTestRunner(tp);
     
             tr.run();
     
             // Assert
             runValidations(() => {
-             assert(tr.failed, "should have failed");
-             assert.strictEqual(tr.warningIssues.length, 0, "should have no warnings");
-             assert.strictEqual(tr.errorIssues.length, 1, "should have 1 error issue");
-             assert(tr.errorIssues[0].startsWith("Error occurred while reading changes folder"), "should have correct message");
+             assert(tr.failed, 'should have failed');
+             assert.strictEqual(tr.warningIssues.length, 0, 'should have no warnings');
+             assert.strictEqual(tr.errorIssues.length, 1, 'should have 1 error issue');
+             assert(tr.errorIssues[0].startsWith('Error occurred while reading changes folder'), 'should have correct message');
             }, 
             tr, done);
         });
@@ -162,7 +162,7 @@ describe("Merge Changelog Test Suite", function() {
         /**
          * Testing the task when valid inputs are passed and everything exists.
          */
-        it("should succeed when everything is at its place", (done: Mocha.Done) => {
+        it('should succeed when everything is at its place', (done: Mocha.Done) => {
             this.timeout(5000);
 
             // Arrange
@@ -170,16 +170,16 @@ describe("Merge Changelog Test Suite", function() {
             createChangelogFile();
 
             // Act
-            let tp: string = path.join(__dirname, "L0Valid.js");
+            let tp: string = path.join(__dirname, 'L0Valid.js');
             let tr: ttm.MockTestRunner = new ttm.MockTestRunner(tp);
 
             tr.run();
 
             // Assert
             runValidations(() => {
-                assert(tr.succeeded, "should have succeeded");
-                assert.strictEqual(tr.warningIssues.length, 0, "should have no warnings");
-                assert.strictEqual(tr.errorIssues.length, 0, "should have 0 error issue");
+                assert(tr.succeeded, 'should have succeeded');
+                assert.strictEqual(tr.warningIssues.length, 0, 'should have no warnings');
+                assert.strictEqual(tr.errorIssues.length, 0, 'should have 0 error issue');
                }, 
                tr, done);
         });
@@ -188,12 +188,12 @@ describe("Merge Changelog Test Suite", function() {
     /**
      * Describes tests with invalid inputs.
      */
-    describe("invalid inputs", function() {
+    describe('invalid inputs', function() {
 
         /**
          * Testing the task when invalid changelog location is passed in.
          */
-        it("should fail if input changelog location does not exist", (done: Mocha.Done) => {
+        it('should fail if input changelog location does not exist', (done: Mocha.Done) => {
             this.timeout(1000);
     
             // Arrange
@@ -201,17 +201,17 @@ describe("Merge Changelog Test Suite", function() {
             createChangelogFile();
     
             // Act 
-            let tp: string = path.join(__dirname, "L0InvalidChangelogLocation.js");
+            let tp: string = path.join(__dirname, 'L0InvalidChangelogLocation.js');
             let tr: ttm.MockTestRunner = new ttm.MockTestRunner(tp);
     
             tr.run();
     
             // Assert
             runValidations(() => {
-             assert(tr.failed, "should have failed");
-             assert.strictEqual(tr.warningIssues.length, 0, "should have no warnings");
-             assert.strictEqual(tr.errorIssues.length, 1, "should have 1 error issue");
-             assert(tr.errorIssues[0].startsWith("Not found changelogLocation"), "should have correct message");
+             assert(tr.failed, 'should have failed');
+             assert.strictEqual(tr.warningIssues.length, 0, 'should have no warnings');
+             assert.strictEqual(tr.errorIssues.length, 1, 'should have 1 error issue');
+             assert(tr.errorIssues[0].startsWith('Not found changelogLocation'), 'should have correct message');
             }, 
             tr, done);
         });
@@ -219,7 +219,7 @@ describe("Merge Changelog Test Suite", function() {
         /**
          * Testing the task when invalid changes location is passed in.
          */
-         it("should fail if input changes location does not exist", (done: Mocha.Done) => {
+         it('should fail if input changes location does not exist', (done: Mocha.Done) => {
             this.timeout(1000);
     
             // Arrange
@@ -227,17 +227,17 @@ describe("Merge Changelog Test Suite", function() {
             createChangelogFile();
     
             // Act 
-            let tp: string = path.join(__dirname, "L0InvalidChangesLocation.js");
+            let tp: string = path.join(__dirname, 'L0InvalidChangesLocation.js');
             let tr: ttm.MockTestRunner = new ttm.MockTestRunner(tp);
     
             tr.run();
     
             // Assert
             runValidations(() => {
-             assert(tr.failed, "should have failed");
-             assert.strictEqual(tr.warningIssues.length, 0, "should have no warnings");
-             assert.strictEqual(tr.errorIssues.length, 1, "should have 1 error issue");
-             assert(tr.errorIssues[0].startsWith("Not found changesLocation"), "should have correct message");
+             assert(tr.failed, 'should have failed');
+             assert.strictEqual(tr.warningIssues.length, 0, 'should have no warnings');
+             assert.strictEqual(tr.errorIssues.length, 1, 'should have 1 error issue');
+             assert(tr.errorIssues[0].startsWith('Not found changesLocation'), 'should have correct message');
             }, 
             tr, done);
         });
@@ -245,7 +245,7 @@ describe("Merge Changelog Test Suite", function() {
         /**
          * Testing the task when changes location is an empty string.
          */
-         it("should fail if input changes location is an empty string", (done: Mocha.Done) => {
+         it('should fail if input changes location is an empty string', (done: Mocha.Done) => {
             this.timeout(1000);
     
             // Arrange
@@ -253,17 +253,17 @@ describe("Merge Changelog Test Suite", function() {
             createChangelogFile();
     
             // Act 
-            let tp: string = path.join(__dirname, "L0EmptyChangesLocation.js");
+            let tp: string = path.join(__dirname, 'L0EmptyChangesLocation.js');
             let tr: ttm.MockTestRunner = new ttm.MockTestRunner(tp);
     
             tr.run();
     
             // Assert
             runValidations(() => {
-             assert(tr.failed, "should have failed");
-             assert.strictEqual(tr.warningIssues.length, 0, "should have no warnings");
-             assert.strictEqual(tr.errorIssues.length, 1, "should have 1 error issue");
-             assert(tr.errorIssues[0].startsWith("Input required"), "should have correct message");
+             assert(tr.failed, 'should have failed');
+             assert.strictEqual(tr.warningIssues.length, 0, 'should have no warnings');
+             assert.strictEqual(tr.errorIssues.length, 1, 'should have 1 error issue');
+             assert(tr.errorIssues[0].startsWith('Input required'), 'should have correct message');
             }, 
             tr, done);
         });