+++ Enterwell.CI.Changelog/MarkdownTextService.cs
@@ -50,26 +50,26 @@ namespace Enterwell.CI.Changelog
             var builder = new StringBuilder();
 
             builder.AppendLine();
-            builder.AppendLine(ToH2($"[{versionInformation.SemanticVersion}] - {DateTime.Now:yyyy-MM-dd}"));
+            builder.AppendLine(this.ToH2($"[{versionInformation.SemanticVersion}] - {DateTime.Now:yyyy-MM-dd}"));
 
             var orderedChanges = versionInformation.Changes.OrderBy(c => c.Key).ToList();
 
             if (orderedChanges.Count == 0)
             {
-                builder.AppendLine(ToUnorderedListItem("No new changes"));
+                builder.AppendLine(this.ToUnorderedListItem("No new changes"));
             }
 
-            for (var changeTypeIndex = 0; changeTypeIndex < orderedChanges.Count; changeTypeIndex++)
+            foreach (var change in orderedChanges)
             {
-                builder.AppendLine(ToH3(orderedChanges[changeTypeIndex].Key));
+                builder.AppendLine(this.ToH3(change.Key));
 
-                foreach (var changeDescription in orderedChanges[changeTypeIndex].Value)
+                foreach (var changeInfo in change.Value)
                 {
-                    builder.AppendLine(ToUnorderedListItem(changeDescription));
+                    builder.AppendLine(this.ToUnorderedListItem(changeInfo.ChangeDescription));
                 }
 
-                // If we are not at the last change, we add a blank line separator. (empty line after last change is not necessary).
-                if (changeTypeIndex != orderedChanges.Count - 1)
+                // If we are not at the last change, we add a blank line separator (empty line after the last change is not necessary).
+                if (orderedChanges.IndexOf(change) != orderedChanges.Count - 1)
                 {
                     builder.AppendLine();
                 }