+++ CHANGELOG.md
@@ -4,16 +4,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2022-05-25
+### Added
+- [CLI] '.changelog.json' file discovery
+- [CoreApp] Opt-in AutoBump option (set the version to appropriate project file)
+- [DevOpsExtension] MergeChangelog@2 Azure Pipelines task that can be ran on both Ubuntu and Windows VMs
+
+### Changed
+- [CLI] Passed in category is now case-insensitive
+- [CLI] Updated README
+- [CoreApp] Only deleting change files that were valid, accepted and used to create a new changelog version
+- [CoreApp] Updated README
+- [DevOpsExtension] Updated README
+
+### Fixed
+- [CLI] Removing excess whitespace from the entered change description and category
+- [VSIX] Removing excess whitespace from the entered change description and category
+
 ## [2.1.0] - 2022-04-11
 ### Added
 - [CoreApp] Outputting newly bumped semantic version
-- [DevOpsTask] Outputting newly bumped semantic version as an output variable called 'bumpedSemanticVersion'
+- [DevOpsExtension] Outputting newly bumped semantic version as an output variable called 'bumpedSemanticVersion'
 
 ## [2.0.0] - 2022-04-07
 ### Added
 - [CoreApp] Automatic version bumping
 - [CoreApp] Defining custom bump rules is available through the configuration
-- [DevOpsTask] MergeChangelog@2 that does not need an explicit semantic version
+- [DevOpsExtension] MergeChangelog@2 that does not need an explicit semantic version
 
 ### Changed
 - [CLI] Using .NET 6
@@ -26,7 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.0] - 2021-09-22
 ### Added
 - [CLI] --version option
-- [CoreApp] 'changes' folder discovery
+- [CLI] 'changes' folder discovery
 - [VSIX] VS 2022 support
 
 ### Fixed
@@ -36,5 +53,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [CLI] Initial version
 - [CoreApp] Core logic implemented
-- [DevOpsTask] Initial version
+- [DevOpsExtension] Initial version
 - [VSIX] Initial version